move fillGap to room
