+++ src/domain/session/room/timeline/tilesCreator.js
@@ -4,11 +4,11 @@ import LocationTile from "./tiles/LocationTile.js";
 import RoomNameTile from "./tiles/RoomNameTile.js";
 import RoomMemberTile from "./tiles/RoomMemberTile.js";
 
-export default function ({timeline, ownUserId}) {
+export default function ({room, ownUserId}) {
     return function tilesCreator(entry, emitUpdate) {
         const options = {entry, emitUpdate, ownUserId};
         if (entry.isGap) {
-            return new GapTile(options, timeline);
+            return new GapTile(options, room);
         } else if (entry.eventType) {
             switch (entry.eventType) {
                 case "m.room.message": {
