+++ src/matrix/room/room.js
@@ -1,6 +1,7 @@
 import EventEmitter from "../../EventEmitter.js";
 import RoomSummary from "./summary.js";
 import SyncWriter from "./timeline/persistence/SyncWriter.js";
+import GapWriter from "./timeline/persistence/GapWriter.js";
 import Timeline from "./timeline/Timeline.js";
 import FragmentIdComparer from "./timeline/FragmentIdComparer.js";
 import SendQueue from "./sending/SendQueue.js";
@@ -58,6 +59,26 @@ export default class Room extends EventEmitter {
         this._sendQueue.enqueueEvent(eventType, content);
     }
 
+
+    /** @public */
+    async fillGap(fragmentEntry, amount) {
+        const response = await this._hsApi.messages(this._roomId, {
+            from: fragmentEntry.token,
+            dir: fragmentEntry.direction.asApiString(),
+            limit: amount,
+            filter: {lazy_load_members: true}
+        }).response();
+        const gapWriter = new GapWriter({
+            roomId: this._roomId,
+            storage: this._storage,
+            fragmentIdComparer: this._fragmentIdComparer
+        });
+        const newEntries = await gapWriter.writeFragmentFill(fragmentEntry, response);
+        if (this._timeline) {
+            this._timeline.addGapEntries(newEntries)
+        }
+    }
+
     get name() {
         return this._summary.name;
     }
@@ -73,7 +94,6 @@ export default class Room extends EventEmitter {
         this._timeline = new Timeline({
             roomId: this.id,
             storage: this._storage,
-            hsApi: this._hsApi,
             fragmentIdComparer: this._fragmentIdComparer,
             pendingEvents: this._sendQueue.pendingEvents,
             closeCallback: () => this._timeline = null,
