+++ src/matrix/room/timeline/Timeline.js
@@ -1,16 +1,14 @@
 import { SortedArray, MappedList, ConcatList } from "../../../observable/index.js";
 import Direction from "./Direction.js";
-import GapWriter from "./persistence/GapWriter.js";
 import TimelineReader from "./persistence/TimelineReader.js";
 import PendingEventEntry from "./entries/PendingEventEntry.js";
 
 export default class Timeline {
-    constructor({roomId, storage, closeCallback, fragmentIdComparer, pendingEvents, user, hsApi}) {
+    constructor({roomId, storage, closeCallback, fragmentIdComparer, pendingEvents, user}) {
         this._roomId = roomId;
         this._storage = storage;
         this._closeCallback = closeCallback;
         this._fragmentIdComparer = fragmentIdComparer;
-        this._hsApi = hsApi;
         this._remoteEntries = new SortedArray((a, b) => a.compare(b));
         this._timelineReader = new TimelineReader({
             roomId: this._roomId,
@@ -36,23 +34,11 @@ export default class Timeline {
         this._remoteEntries.setManySorted(newEntries);
     }
 
-    /** @public */
-    async fillGap(fragmentEntry, amount) {
-        const response = await this._hsApi.messages(this._roomId, {
-            from: fragmentEntry.token,
-            dir: fragmentEntry.direction.asApiString(),
-            limit: amount,
-            filter: {lazy_load_members: true}
-        }).response();
-        const gapWriter = new GapWriter({
-            roomId: this._roomId,
-            storage: this._storage,
-            fragmentIdComparer: this._fragmentIdComparer
-        });
-        const newEntries = await gapWriter.writeFragmentFill(fragmentEntry, response);
+    /** @package */
+    addGapEntries(newEntries) {
         this._remoteEntries.setManySorted(newEntries);
     }
-
+    
     // tries to prepend `amount` entries to the `entries` list.
     async loadAtTop(amount) {
         const firstEventEntry = this._remoteEntries.array.find(e => !!e.eventType);
